requirements.txt
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-configurations==2.0
django-friendship==1.5.0
django-nose==1.4.5
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
lxml==4.1.1
nose==1.3.7
oauthlib==2.0.6
parso==0.1.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
